refactor(Badge): move component to content folder
